requirements.txt
Django>=4.2.0
django-filter>=23.5
django-cors-headers>=4.3.0
djangorestframework>=3.14.0
graphene-django>=3.1.5
transformers>=4.36.0
torch>=2.1.0
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0